Speed up ANALYZE by computing column-level statistics in memory.
Previously ANALYZE in GPDB uses up to 6 SPI calls per column to compute
column level statistics including null fraction, average width,
ndistinct, mcv and histogram. That resulted in slow performance when the
number of columns and number of partitions are large.

In this overhaul of Analyze, we keep the collection of table level stats
(reltuples, relpages) and the sampling mechanism unchanged. For
column-level stats, we use a similar approach as PostgreSQL. We also
have some optimization for string typed columns since they are slow to
sort in a plain way. We leverage the ability of mk_sort in the executor
to sort string datums much efficiently.